moved chord voicer to instruments
