+++ src/instruments/piano/chord_voicer.py
@@ -3,7 +3,7 @@ from typing import FrozenSet, List
 from common.structures.chord import ChordQuality
 from common.structures.pitch import Pitch
 
-from generation.chord_voicer import (
+from instruments.chord_voicer import (
     ChordVoicer,
     ChordInChordProgression,
     ChordVoicerMemo,
